Rename package and Params type
